Restructuring generics out of sign-in-side
