+++ frontend/theme/customizations/dataDisplay.tsx
